+++ src/k8secrets/__main__.py
@@ -1,15 +1,19 @@
 import argparse
 import sys
 
-from .secrets import create_env_map, create_secret_object, parse
-
+from .secrets import create_config_object, create_env_map, create_secret_object, parse
 
 DESCRIPTION = """\
 description:
 
-'k8secrets' takes a secret object name and a list of variables as input:
+'k8secrets' takes an object name and a list of variables as input and creates a
+"Secret" Kubernetes object:
 
     k8secrets mysecret -l KEY1=value1,KEY2=value2
+
+or a "ConfigMap":
+
+    k8secrets --config myconfig -l KEY1=value1,KEY2=value2
 
 
 It can also read the variables from stdin (you can pipe input into it):
@@ -28,7 +32,7 @@ be separated either by a newline character (Unix or Windows) or a comma or a
 semicolon.
 
 'k8secrets' prints output to stdout. The output is valid yaml and contains
-two sections. First is a complete k8s secret object:
+two sections. First is a complete k8s Secret of ConfigMap object:
 
     ---
     apiVersion: v1
@@ -40,35 +44,28 @@ two sections. First is a complete k8s secret object:
       key1: dmFsdWUx
       key2: dmFsdWUy
 
-The second is an 'env' section with references to the secret object that you
-can use in your deployments or cronjobs:
+The second is an 'envFrom' section with a reference to the Secret or ConfigMap
+object that you can use in your deployments or cronjobs:
 
     ---
-    env:
-      - name: KEY1
-        valueFrom:
-          secretKeyRef:
-            name: mysecret
-            key: key1
-      - name: KEY2
-        valueFrom:
-          secretKeyRef:
-            name: mysecret
-            key: key2
+    envFrom:
+      - secretRef:
+          name: mysecret
 """
 
 
 PARSER = argparse.ArgumentParser(
     description=DESCRIPTION, formatter_class=argparse.RawDescriptionHelpFormatter
 )
-PARSER.add_argument("secret_name", type=str)
+PARSER.add_argument("--config", action="store_true", help="Create a ConfigMap object")
+PARSER.add_argument("object_name", type=str)
 PARSER.add_argument("-l", type=str, required=False, dest="variable_list")
 
 
 def main():
 
     args = PARSER.parse_args()
-    secret_name = args.secret_name
+    object_name = args.object_name
     variable_list = args.variable_list
 
     if not variable_list and not sys.stdin.isatty():
@@ -81,9 +78,12 @@ def main():
     variables = parse(variable_list)
 
     print()
-    print(create_secret_object(secret_name, variables))
+    if args.config:
+        print(create_config_object(object_name, variables))
+    else:
+        print(create_secret_object(object_name, variables))
     print()
-    print(create_env_map(secret_name, variables))
+    print(create_env_map(object_name))
 
 
 if __name__ == "__main__":