+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog #
 
+## Version 1.0.0 ##
+Changes:
+- Use `poetry`
+- Add `--config` flag to generate `ConfigMaps`
+- Generate an `envFrom` section instead of listing all variables
+
 ## Version 0.0.1 ##
 Features:
 - Implementation